fixes to login register

+++ backend/src/services/design-service.ts
@@ -18,7 +18,7 @@ export default class DesignService {
     return design.id;
   }
   public async delete(user: User, id: string): Promise<string> {
-    this.checkDesign(user, id);
+    this.authorize(user, id);
     await this.prisma.design.delete({
       where: {
         id,
@@ -31,9 +31,8 @@ export default class DesignService {
     user: User,
     id: string,
     objects: any
-): Promise<string> {
-
-    this.checkDesign(user, id);
+  ): Promise<string> {
+    this.authorize(user, id);
 
     const objList = await this.prisma.objects.findMany({
       where: {
@@ -57,27 +56,123 @@ export default class DesignService {
         await this.prisma.objects.create({
           data: {
             id: obj.id,
-            layerIndex: obj.layerIndex as number,  // Assuming `obj` has the `layerIndex` property
+            layerIndex: obj.layerIndex as number,
             data: JSON.stringify(obj),
             designId: id,
           },
         });
-        
       }
     }
 
-    // (Optional) If you want to handle deletions, you'd also loop through objList 
-    // and check if any object there doesn't exist in the `objects` array and then remove it.
-
     return "updated";
-}
+  }
+
+  public async moveUp(
+    user: User,
+    designId: string,
+    id: string
+  ): Promise<string> {
+    this.authorize(user, id);
+
+    const obj = await this.prisma.objects.findFirst({
+      where: {
+        id: id,
+        designId: designId,
+      },
+    });
+
+    if (!obj) {
+      throw new Error("No object was found");
+    }
+
+    const objAbove = await this.prisma.objects.findFirst({
+      where: {
+        designId: designId,
+        layerIndex: obj.layerIndex - 1,
+      },
+    });
+
+    if (!objAbove) {
+      throw new Error("No object above");
+    }
+
+    await this.prisma.objects.update({
+      where: {
+        id: obj.id,
+      },
+      data: {
+        layerIndex: obj.layerIndex - 1,
+      },
+    });
+
+    await this.prisma.objects.update({
+      where: {
+        id: objAbove.id,
+      },
+      data: {
+        layerIndex: objAbove.layerIndex + 1,
+      },
+    });
+
+    return "moved up";
+  }
+
+  public async moveDown(
+    user: User,
+    designId: string,
+    id: string
+  ): Promise<string> {
+    this.authorize(user, designId);
+
+    const obj = await this.prisma.objects.findFirst({
+      where: {
+        id: id,
+        designId: designId,
+      },
+    });
+
+    if (!obj) {
+      throw new Error("No object was found");
+    }
+
+    const objBelow = await this.prisma.objects.findFirst({
+      where: {
+        designId: designId,
+        layerIndex: obj.layerIndex + 1,
+      },
+    });
+
+    if (!objBelow) {
+      throw new Error("No object below");
+    }
+
+    await this.prisma.objects.update({
+      where: {
+        id: obj.id,
+      },
+      data: {
+        layerIndex: obj.layerIndex + 1,
+      },
+    });
+
+    await this.prisma.objects.update({
+      where: {
+        id: objBelow.id,
+      },
+      data: {
+        layerIndex: objBelow.layerIndex - 1,
+      },
+    });
+
+    return "moved down";
+  }
 
   public async getAll(user: User): Promise<string> {
     const designs = await this.prisma.design.findMany();
     return JSON.stringify(designs);
   }
 
-  private async checkDesign(user: User, id: string) {
+  private async authorize(user: User, id: string) {
     const design = await this.prisma.design.findFirst({
       where: {
         id: id,
